refactor: move cards container to separate file
